gateway: split local osmosis swap planner services
